+++ packages/theme/terrazzo.config.ts
@@ -156,7 +156,7 @@ const config: Config = {
 						'Background color variants for surface elements.',
 					patterns: [
 						{
-							pattern: /^wpds-color\.bg\.surface\.(.+)$/,
+							pattern: /^wpds-color\.background\.surface\.(.+)$/,
 							transform: ( variant ) =>
 								variant.split( '.' ).join( '-' ),
 						},
@@ -168,7 +168,8 @@ const config: Config = {
 						'Background color variants for interactive elements.',
 					patterns: [
 						{
-							pattern: /^wpds-color\.bg\.interactive\.(.+)$/,
+							pattern:
+								/^wpds-color\.background\.interactive\.(.+)$/,
 							transform: ( variant ) =>
 								variant
 									.split( '.' )
@@ -183,7 +184,7 @@ const config: Config = {
 						'Foreground color variants for content text and icons.',
 					patterns: [
 						{
-							pattern: /^wpds-color\.fg\.content\.(.+)$/,
+							pattern: /^wpds-color\.foreground\.content\.(.+)$/,
 							transform: ( variant ) =>
 								variant.split( '.' ).join( '-' ),
 						},
@@ -195,7 +196,8 @@ const config: Config = {
 						'Foreground color variants for interactive element text and icons.',
 					patterns: [
 						{
-							pattern: /^wpds-color\.fg\.interactive\.(.+)$/,
+							pattern:
+								/^wpds-color\.foreground\.interactive\.(.+)$/,
 							transform: ( variant ) =>
 								variant
 									.split( '.' )
@@ -235,7 +237,7 @@ const config: Config = {
 					description: 'Foreground color variants for text elements.',
 					patterns: [
 						{
-							pattern: /^wpds-color\.fg\.[^.]+\.(.+)$/,
+							pattern: /^wpds-color\.foreground\.[^.]+\.(.+)$/,
 							transform: ( variant ) =>
 								variant.split( '.' ).join( '-' ),
 						},